decompose view.ts
