+++ skymp5-client/src/formView.ts
@@ -1,0 +1,493 @@
+import { Actor, ActorBase, createText, destroyText, Form, Game, NetImmerse, ObjectReference, once, printConsole, setTextPos, setTextString, TESModPlatform, worldPointToScreenPoint } from "skyrimPlatform";
+import { setDefaultAnimsDisabled, applyAnimation } from "./animation";
+import { Appearance, applyAppearance } from "./appearance";
+import { isBadMenuShown, applyEquipment } from "./equipment";
+import { RespawnNeededError } from "./errors";
+import { applyInventory } from "./inventory";
+import { FormModel } from "./model";
+import { applyMovement } from "./movementApply";
+import { getScreenResolution } from "./skyrimSettings";
+import { SpawnProcess } from "./spawnProcess";
+import { UtilsObjectReference } from "./utilsObjectReference";
+import { View } from "./view";
+import { modWcProtection } from "./worldCleaner";
+import * as deathSystem from "./deathSystem";
+import { GamemodeApiSupport } from "./gamemodeApiSupport";
+import { PlayerCharacterDataHolder } from "./playerCharacterDataHolder";
+import { getMovement } from "./movementGet";
+import { lastTryHost, tryHost } from "./hostAttempts";
+
+export class FormView implements View<FormModel> {
+  constructor(private remoteRefrId?: number) { }
+
+  update(model: FormModel): void {
+    // Other players mutate into PC clones when moving to another location
+    if (model.movement) {
+      if (!this.lastWorldOrCell)
+        this.lastWorldOrCell = model.movement.worldOrCell;
+      if (this.lastWorldOrCell !== model.movement.worldOrCell) {
+        printConsole(
+          `[1] worldOrCell changed, destroying FormView ${this.lastWorldOrCell.toString(
+            16
+          )} => ${model.movement.worldOrCell.toString(16)}`
+        );
+        this.lastWorldOrCell = model.movement.worldOrCell;
+        this.destroy();
+        this.refrId = 0;
+        this.appearanceBasedBaseId = 0;
+        return;
+      }
+    }
+
+    // Players with different worldOrCell should be invisible
+    if (model.movement) {
+      const worldOrCell =
+        (Game.getPlayer() as Actor).getWorldSpace() ||
+        (Game.getPlayer() as Actor).getParentCell();
+      if (
+        worldOrCell &&
+        model.movement.worldOrCell !== worldOrCell.getFormID()
+      ) {
+        this.destroy();
+        this.refrId = 0;
+        return;
+      }
+    }
+
+    // Apply appearance before base form selection to prevent double-spawn
+    if (model.appearance) {
+      if (
+        !this.appearanceState.appearance ||
+        model.numAppearanceChanges !== this.appearanceState.lastNumChanges
+      ) {
+        this.appearanceState.appearance = model.appearance;
+        this.appearanceState.lastNumChanges = model.numAppearanceChanges as number;
+        this.appearanceBasedBaseId = 0;
+      }
+    }
+
+    const refId =
+      model.refrId && model.refrId < 0xff000000 ? model.refrId : undefined;
+    if (refId) {
+      if (this.refrId !== refId) {
+        this.destroy();
+        this.refrId = model.refrId as number;
+        this.ready = true;
+        const refr = ObjectReference.from(Game.getFormEx(this.refrId));
+        if (refr) {
+          const base = refr.getBaseObject();
+          if (base) UtilsObjectReference.dealWithRef(refr, base);
+        }
+      }
+    } else {
+      const base =
+        Game.getFormEx(+(model.baseId as number)) ||
+        Game.getFormEx(this.getAppearanceBasedBase());
+      if (!base) return;
+
+      let refr = ObjectReference.from(Game.getFormEx(this.refrId));
+      const respawnRequired =
+        !refr ||
+        !refr.getBaseObject() ||
+        (refr.getBaseObject() as Form).getFormID() !== base.getFormID();
+
+      if (respawnRequired) {
+        this.destroy();
+        refr = (Game.getPlayer() as Actor).placeAtMe(
+          base,
+          1,
+          true,
+          true
+        ) as ObjectReference;
+        this.state = {};
+        delete this.wasHostedByOther;
+        const kTypeNpc = 43;
+        if (base.getType() !== kTypeNpc) {
+          refr.setAngle(
+            model.movement?.rot[0] || 0,
+            model.movement?.rot[1] || 0,
+            model.movement?.rot[2] || 0
+          );
+        }
+        modWcProtection(refr.getFormID(), 1);
+
+        // TODO: reset all states?
+        this.eqState = this.getDefaultEquipState();
+
+        this.ready = false;
+        new SpawnProcess(
+          this.appearanceState.appearance as Appearance,
+          model.movement
+            ? model.movement.pos
+            : UtilsObjectReference.getPos(Game.getPlayer() as Actor),
+          refr.getFormID(),
+          () => {
+            this.ready = true;
+            this.spawnMoment = Date.now();
+          }
+        );
+        if (model.appearance && model.appearance.name) {
+          refr.setDisplayName("" + model.appearance.name, true);
+        }
+        Actor.from(refr)?.setActorValue("attackDamageMult", 0);
+      }
+      this.refrId = (refr as ObjectReference).getFormID();
+    }
+
+    if (!this.ready) return;
+
+    const refr = ObjectReference.from(Game.getFormEx(this.refrId));
+    if (refr) {
+      const actor = Actor.from(refr);
+      if (actor && !this.localImmortal) {
+        deathSystem.makeActorImmortal(actor);
+        actor.setActorValue("health", 1000000);
+        this.localImmortal = true;
+      }
+      this.applyAll(refr, model);
+      GamemodeApiSupport.updateNeighbor(refr, model, this.state);
+    }
+  }
+
+  destroy(): void {
+    this.isOnScreen = false;
+    this.spawnMoment = 0;
+    const refrId = this.refrId;
+    once("update", () => {
+      if (refrId >= 0xff000000) {
+        const refr = ObjectReference.from(Game.getFormEx(refrId));
+        if (refr) refr.delete();
+        modWcProtection(refrId, -1);
+        const ac = Actor.from(refr);
+        if (ac) {
+          TESModPlatform.setWeaponDrawnMode(ac, -1);
+        }
+      }
+    })
+
+    this.localImmortal = false;
+    this.removeNickname();
+  }
+
+  private applyHarvested(refr: ObjectReference, isHarvested: boolean) {
+    const base = refr.getBaseObject();
+    if (base) {
+      const t = base.getType();
+      if (t >= 38 && t <= 39) {
+        const wasHarvested = refr.isHarvested();
+        if (isHarvested != wasHarvested) {
+          let ac: Actor = undefined as unknown as Actor;
+          if (isHarvested)
+            for (let i = 0; i < 20; ++i) {
+              ac = Game.findRandomActor(
+                refr.getPositionX(),
+                refr.getPositionY(),
+                refr.getPositionZ(),
+                10000
+              ) as Actor;
+              if (ac && ac.getFormID() !== 0x14) {
+                break;
+              }
+            }
+          if (isHarvested && ac && ac.getFormID() !== 0x14) {
+            refr.activate(ac, true);
+          } else {
+            refr.setHarvested(isHarvested);
+            const id = refr.getFormID();
+            refr.disable(false).then(() => {
+              const restoredRefr = ObjectReference.from(Game.getFormEx(id));
+              if (restoredRefr) restoredRefr.enable(false);
+            });
+          }
+        }
+      } else {
+        const wasHarvested = refr.isDisabled();
+        if (isHarvested != wasHarvested) {
+          if (isHarvested) {
+            const id = refr.getFormID();
+            refr.disable(false).then(() => {
+              const restoredRefr = ObjectReference.from(Game.getFormEx(id));
+              if (restoredRefr && !restoredRefr.isDisabled()) {
+                restoredRefr.delete();
+                // Deletion takes time, so in practice this would be called a lot of times
+              }
+            });
+          } else refr.enable(true);
+        }
+      }
+    }
+  }
+
+  private lastHarvestedApply = 0;
+  private lastOpenApply = 0;
+
+  private applyAll(refr: ObjectReference, model: FormModel) {
+    let forcedWeapDrawn: boolean | null = null;
+
+    if (PlayerCharacterDataHolder.getCrosshairRefId() === this.refrId) {
+      this.lastHarvestedApply = 0;
+      this.lastOpenApply = 0;
+    }
+    const now = Date.now();
+    if (now - this.lastHarvestedApply > 666) {
+      this.lastHarvestedApply = now;
+      this.applyHarvested(refr, !!model.isHarvested);
+    }
+    if (now - this.lastOpenApply > 133) {
+      this.lastOpenApply = now;
+      refr.setOpen(!!model.isOpen);
+    }
+
+    if (
+      model.inventory &&
+      PlayerCharacterDataHolder.getCrosshairRefId() == this.refrId &&
+      !isBadMenuShown()
+    ) {
+      // Do not let actors breaking their equipment via inventory apply
+      // However, actually, actors do not have inventory in their models
+      // Except your clone.
+      if (!Actor.from(refr)) {
+        applyInventory(refr, model.inventory, false, true);
+      }
+    }
+
+    if (model.animation) {
+      if (model.animation.animEventName === "SkympFakeUnequip") {
+        forcedWeapDrawn = false;
+      } else if (model.animation.animEventName === "SkympFakeEquip") {
+        forcedWeapDrawn = true;
+      }
+    }
+
+    if (model.movement) {
+      const ac = Actor.from(refr);
+      if (ac) {
+        if (model.isHostedByOther !== this.wasHostedByOther) {
+          this.wasHostedByOther = model.isHostedByOther;
+          this.movState.lastApply = 0;
+          if (model.isHostedByOther) {
+            setDefaultAnimsDisabled(ac.getFormID(), true);
+          } else {
+            setDefaultAnimsDisabled(ac.getFormID(), false);
+          }
+        }
+      }
+
+      if (
+        this.movState.lastApply &&
+        Date.now() - this.movState.lastApply > 1500
+      ) {
+        if (Date.now() - this.movState.lastRehost > 1000) {
+          this.movState.lastRehost = Date.now();
+          const remoteId = this.remoteRefrId;
+          if (ac && ac.is3DLoaded()) {
+            this.tryHostIfNeed(ac, remoteId as number);
+          }
+        }
+      }
+
+      if (
+        +(model.numMovementChanges as number) !==
+        this.movState.lastNumChanges ||
+        Date.now() - this.movState.lastApply > 2000
+      ) {
+        this.movState.lastApply = Date.now();
+        if (model.isHostedByOther || !this.movState.everApplied) {
+          const backup = model.movement.isWeapDrawn;
+          if (forcedWeapDrawn === true || forcedWeapDrawn === false) {
+            model.movement.isWeapDrawn = forcedWeapDrawn;
+          }
+          try {
+            applyMovement(refr, model.movement, !!model.isMyClone);
+          } catch (e) {
+            if (e instanceof RespawnNeededError) {
+              this.lastWorldOrCell = model.movement.worldOrCell;
+              this.destroy();
+              this.refrId = 0;
+              this.appearanceBasedBaseId = 0;
+              return;
+            } else {
+              throw e;
+            }
+          }
+          model.movement.isWeapDrawn = backup;
+
+          this.movState.lastNumChanges = +(model.numMovementChanges as number);
+          this.movState.everApplied = true;
+        } else {
+          if (ac) {
+            ac.clearKeepOffsetFromActor(); 
+            TESModPlatform.setWeaponDrawnMode(ac, -1);
+          }
+          const remoteId = this.remoteRefrId;
+          if (ac && remoteId && ac.is3DLoaded()) {
+            this.tryHostIfNeed(ac, remoteId);
+          }
+        }
+      }
+    }
+    if (model.animation) applyAnimation(refr, model.animation, this.animState);
+
+    if (model.appearance) {
+      const actor = Actor.from(refr);
+      if (actor && !PlayerCharacterDataHolder.isInJumpState()) {
+        if (PlayerCharacterDataHolder.getWorldOrCell()) {
+          if (
+            this.lastPcWorldOrCell &&
+            PlayerCharacterDataHolder.getWorldOrCell() !== this.lastPcWorldOrCell
+          ) {
+            // Redraw tints if PC world/cell changed
+            this.isOnScreen = false;
+          }
+          this.lastPcWorldOrCell = PlayerCharacterDataHolder.getWorldOrCell();
+        }
+
+        const headPos = [
+          NetImmerse.getNodeWorldPositionX(actor, "NPC Head [Head]", false),
+          NetImmerse.getNodeWorldPositionY(actor, "NPC Head [Head]", false),
+          NetImmerse.getNodeWorldPositionZ(actor, "NPC Head [Head]", false),
+        ];
+        const [screenPoint] = worldPointToScreenPoint(headPos);
+        const isOnScreen =
+          screenPoint[0] > 0 &&
+          screenPoint[1] > 0 &&
+          screenPoint[2] > 0 &&
+          screenPoint[0] < 1 &&
+          screenPoint[1] < 1 &&
+          screenPoint[2] < 1;
+        if (isOnScreen != this.isOnScreen) {
+          this.isOnScreen = isOnScreen;
+          if (isOnScreen) {
+            actor.queueNiNodeUpdate();
+            (Game.getPlayer() as Actor).queueNiNodeUpdate();
+          }
+        }
+      }
+    }
+
+    if (model.equipment) {
+      const isShown = isBadMenuShown();
+      if (this.eqState.isBadMenuShown !== isShown) {
+        this.eqState.isBadMenuShown = isShown;
+        if (!isShown) this.eqState.lastNumChanges = -1;
+      }
+      if (this.eqState.lastNumChanges !== model.equipment.numChanges) {
+        const ac = Actor.from(refr);
+        // If we do not block inventory here, we will be able to reproduce the bug:
+        // 1. Place ~90 bots and force them to reequip iron swords to the left hand (rate should be ~50ms)
+        // 2. Open your inventory and reequip different items fast
+        // 3. After 1-2 minutes close your inventory and see that HUD disappeared
+        if (
+          ac &&
+          !isBadMenuShown() &&
+          Date.now() - this.eqState.lastEqMoment > 500 &&
+          Date.now() - this.spawnMoment > -1 &&
+          this.spawnMoment > 0
+        ) {
+          //if (this.spawnMoment > 0 && Date.now() - this.spawnMoment > 5000) {
+          if (applyEquipment(ac, model.equipment)) {
+            this.eqState.lastNumChanges = model.equipment.numChanges;
+          }
+          this.eqState.lastEqMoment = Date.now();
+          //}
+          //const res: boolean = applyEquipment(ac, model.equipment);
+          //if (res) this.eqState.lastNumChanges = model.equipment.numChanges;
+        }
+      }
+    }
+
+    if (this.refrId && model.appearance?.name) {
+      const headPart = "NPC Head [Head]";
+      const maxNicknameDrawDistance = 1000;
+      const playerActor = Game.getPlayer()!;
+      const isVisibleByPlayer = !model.movement?.isSneaking && playerActor.getDistance(refr) <= maxNicknameDrawDistance && playerActor.hasLOS(refr);
+      if (isVisibleByPlayer) {
+        const headScreenPos = worldPointToScreenPoint([
+          NetImmerse.getNodeWorldPositionX(refr, headPart, false),
+          NetImmerse.getNodeWorldPositionY(refr, headPart, false),
+          NetImmerse.getNodeWorldPositionZ(refr, headPart, false) + 22
+        ])[0];
+        const resolution = getScreenResolution();
+        const textXPos = Math.round(headScreenPos[0] * resolution.width);
+        const textYPos = Math.round((1 - headScreenPos[1]) * resolution.height);
+
+        if (!this.textNameId) {
+          this.textNameId = createText(textXPos, textYPos, model.appearance.name, [255, 255, 255, 1]);
+        } else {
+          setTextString(this.textNameId, headScreenPos[2] >= 0 ? model.appearance.name : "");
+          setTextPos(this.textNameId, textXPos, textYPos);
+        }
+      } else {
+        this.removeNickname();
+      }
+    } else {
+      this.removeNickname();
+    }
+  }
+
+  private removeNickname() {
+    if (this.textNameId) {
+      destroyText(this.textNameId);
+      this.textNameId = undefined;
+    }
+  }
+
+  private getAppearanceBasedBase(): number {
+    const base = ActorBase.from(Game.getFormEx(this.appearanceBasedBaseId));
+    if (!base && this.appearanceState.appearance) {
+      this.appearanceBasedBaseId = applyAppearance(this.appearanceState.appearance).getFormID();
+    }
+    return this.appearanceBasedBaseId;
+  }
+
+  private getDefaultEquipState() {
+    return { lastNumChanges: 0, isBadMenuShown: false, lastEqMoment: 0 };
+  };
+
+  private getDefaultAppearanceState() {
+    return { lastNumChanges: 0, appearance: null as (null | Appearance) };
+  };
+
+  private tryHostIfNeed(ac: Actor, remoteId: number) {
+    const last = lastTryHost[remoteId];
+    if (!last || Date.now() - last >= 1000) {
+      lastTryHost[remoteId] = Date.now();
+  
+      if (
+        getMovement(ac).worldOrCell ===
+        getMovement(Game.getPlayer() as Actor).worldOrCell
+      ) {
+        return tryHost(remoteId);
+      }
+    }
+  };
+
+  getLocalRefrId(): number {
+    return this.refrId;
+  }
+
+  getRemoteRefrId(): number {
+    return this.remoteRefrId as number;
+  }
+
+  private refrId = 0;
+  private ready = false;
+  private animState = { lastNumChanges: 0 };
+  private movState = {
+    lastNumChanges: 0,
+    lastApply: 0,
+    lastRehost: 0,
+    everApplied: false,
+  };
+  private appearanceState = this.getDefaultAppearanceState();
+  private eqState = this.getDefaultEquipState();
+  private appearanceBasedBaseId = 0;
+  private isOnScreen = false;
+  private lastPcWorldOrCell = 0;
+  private lastWorldOrCell = 0;
+  private spawnMoment = 0;
+  private wasHostedByOther: boolean | undefined = undefined;
+  private state = {};
+  private localImmortal = false;
+  private textNameId: number | undefined = undefined;
+}
